+++ cupy/cuda/cupy_cub.h
@@ -28,13 +28,13 @@
 void cub_device_reduce(void*, size_t&, void*, void*, size_t, cudaStream_t, int, int);
 void cub_device_segmented_reduce(void*, size_t&, void*, void*, size_t, size_t, cudaStream_t, int, int);
 void cub_device_scan(void*, size_t&, void*, void*, size_t, cudaStream_t, int, int);
-void cub_device_histogram_range(void*, size_t&, void*, void*, size_t, void*, size_t, cudaStream_t, int);
-void cub_device_histogram_even(void*, size_t&, void*, void*, size_t, size_t, size_t, size_t, cudaStream_t, int);
+void cub_device_histogram_range(void*, size_t&, void*, void*, int, void*, size_t, cudaStream_t, int);
+void cub_device_histogram_even(void*, size_t&, void*, void*, int, int, int, size_t, cudaStream_t, int);
 size_t cub_device_reduce_get_workspace_size(void*, void*, size_t, cudaStream_t, int, int);
 size_t cub_device_segmented_reduce_get_workspace_size(void*, void*, size_t, size_t, cudaStream_t, int, int);
 size_t cub_device_scan_get_workspace_size(void*, void*, size_t, cudaStream_t, int, int);
-size_t cub_device_histogram_range_get_workspace_size(void*, void*, size_t, void*, size_t, cudaStream_t, int);
-size_t cub_device_histogram_even_get_workspace_size(void*, void*, size_t, size_t, size_t, size_t, cudaStream_t, int);
+size_t cub_device_histogram_range_get_workspace_size(void*, void*, int, void*, size_t, cudaStream_t, int);
+size_t cub_device_histogram_even_get_workspace_size(void*, void*, int, int, int, size_t, cudaStream_t, int);
 
 // This is for CUB's HistogramRange; hipCUB does not need this (see comment in cupy_cub.cu)
 #ifdef __CUDA_ARCH__
